pull modal-dialog container into the container folder to be "pure" react
