+++ src/browser/jsx/containers/modal-dialog-viewer/modal-dialog.actions.js
