+++ src/browser/jsx/containers/modal-dialog-viewer/modal-dialog.reducer.js
